fix(7a): tighten assessment — integer scores, 3-4 task counts, sonnet-4-6
Verified end-to-end against Claude — all 6 schema constraints pass on
the test call (scoreToday/Projected/WithPlan integer, all 5 factor
scores integer, exposedTasks and defensibleTasks each 3-4 items).

Model swap:
- lib/anthropic.ts: claude-sonnet-4-20250514 -> claude-sonnet-4-6.
  The previous snapshot was deprecated, EOL 2026-06-15. Same Sonnet
  tier; observed output quality slightly higher and more specific.
- docs/kickoff.md §3 updated to match (pinned snapshot, not alias).

System prompt (in lib/anthropic.ts and verbatim in docs/kickoff.md §6):
- Added explicit "all scores MUST be whole-number integers 0-10, no
  decimals" instruction. The JSON template's comment alone was
  insufficient — Claude returned 7.2, 8.4, etc.
- Added explicit "exposedTasks MUST contain exactly 3 to 4 items"
  and same for defensibleTasks. Same root cause: JSON-comment-only
  rules were ignored; Claude returned 7 items each.

Zod (app/api/assess-exposure/route.ts):
- Added .int() to scoreToday, scoreProjected, scoreWithPlan, and
  factor.score. Defence-in-depth alongside the prompt instructions.

scripts/test-assess.mjs:
- Bumped model + prompt to match lib/anthropic.ts (kept in sync by
  hand; throwaway script, not worth a TS-import shim).
- Added inline constraint checks at the end so future re-runs show
  pass/fail at a glance without manual JSON inspection.

cache_control on the system message: left as-is — still a no-op below
Anthropic's ~1024-token caching threshold. Harmless until the prompt
grows past the threshold.

+++ app/api/assess-exposure/route.ts
@@ -33,15 +33,15 @@ const RequestSchema = z.object({
 
 const FactorSchema = z.object({
   label: z.enum(FIVE_FACTOR_LABELS),
-  score: z.number().min(0).max(10),
+  score: z.number().int().min(0).max(10),
   note: z.string().min(1),
 });
 
 const AssessmentSchema = z.object({
   occupationLabel: z.string().min(1),
-  scoreToday: z.number().min(0).max(10),
-  scoreProjected: z.number().min(0).max(10),
-  scoreWithPlan: z.number().min(0).max(10),
+  scoreToday: z.number().int().min(0).max(10),
+  scoreProjected: z.number().int().min(0).max(10),
+  scoreWithPlan: z.number().int().min(0).max(10),
   factors: z.array(FactorSchema).length(5),
   exposedTasks: z.array(z.string().min(1)).min(3).max(4),
   defensibleTasks: z.array(z.string().min(1)).min(3).max(4),
